pyproject.toml
[project]
name = "wispo-roboto"
version = "1.0.0"
description = "Telegram bot for WISPO group"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-telegram-bot>=20.7",
    "requests>=2.31.0",
    "geopy>=2.4.1",
    "google-cloud-secret-manager>=2.18.0",
    "google-cloud-firestore>=2.14.0",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
